add bash autocomplete for equery commands
With this CL, at the chroot prompt if you enter
'equery [tab][tab]' then bash will show you an autocomplete list of
equery commands.

I'm not 100% sure this is the right place to add this funcionality,
since most of the other progammable autocomplete stuff in here is for
cros specific tools. But it is a nice convenience.

BUG=None
TEST=enter chroot. type 'equery [tab][tab]'. See list of equery
commands.

Change-Id: I186e14ac4b9e823885c0a7c7f4177603cbf3e682
Reviewed-on: https://gerrit.chromium.org/gerrit/45247
Reviewed-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Darin Petkov <petkov@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>