Stop calling generate_au_zip inside build_image.
This is part of moving the call to generate_au_zip.py into cbuildbot and
out of the general build_image. We don't need that script if we aren't
building on the servers, and were even explicitly continuing if it failed.

So... stop doing it in every build, and only do it on the builders which
is when it's really needed.

BUG=chromium-os:214786
TEST=Built, build_image, image_to_live.sh.
CQ-DEPEND=Ibfea3c0c4d09235d24a590e53acb4e6506d145cf

Change-Id: Ibc39d15ef4dfa30b82330432d8b10d8154ba7545
Reviewed-on: https://gerrit.chromium.org/gerrit/45953
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>