update_kernel: add option to disable syslinux kernel update
This makes it easier to do update_kernel to a system that has booted
without modules, since the syslinux partition tends to be FAT formatted
and we don't have that statically built into the kernel most of the time.

The code has just been moved (with whitespace reformatting).

BUG=none
TEST=adhoc; make sure update_kernel still works and that --nosyslinux
skips syslinux kernel install.

Change-Id: I07a4c6a240ef57846a7f8c7828ef17583b03b7b6
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/48949
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>