Remove obselete VM disk layout.
Use case is either for an updatable vm layout where users should use
2gb-rootfs-updatable

OR

usb.

This reduces the size of the default VM image to 2.4GB back to where it was
before some of the disk layout refactoring.

In this CL I also cleanup the now unused VM constants that specified partition
numbers etc.

BUG=chromium:234556
TEST=canary + paladin trybot run.

Change-Id: I197dba95c6507edfe3f38ff8120be64a380691a8
Reviewed-on: https://gerrit.chromium.org/gerrit/48962
Commit-Queue: Chris Sosa <sosa@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Peter Mayo <petermayo@chromium.org>