cgpt: improve blank partition handling
We want to be able to omit certain partitions when they don't make sense.
Like throwing away the EFI partition on a device that doesn't have an EFI
bootloader (or anything like it).

Currently if we try to do that in the disk layout (by setting them as
blank), the build scripts barf because cgpt returns an error when they
request information by partition #.  Now we return the right info, and
then the build scripts will use that to figure out what it should do.

We also fix the main partition write script to skip blank partitions.

BUG=chromium:236941
TEST=`./build_image` w/json declaring a blank EFI partition produces an image w/out EFI
TEST=`cbuildbot {amd64,x86,arm}-generic-full` works

Change-Id: I009c1eb2973fe18081432d019a008529adaff4ae
Reviewed-on: https://gerrit.chromium.org/gerrit/59845
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>