cgpt: allow sub-configs to inherit from base
The vast majority of the time, we want to tweak the size of partitions in
the sub-configs only.  Rather than requiring people to duplicate all the
other fields, have the config automatically populate the fields from the
settings found in the base config.

Now people need only specify the partition num to override an earlier one.
This will be more important in follow up commits that add more fields.

BUG=chromium:236941
TEST=`./cgpt.py debug usb legacy_disk_layout.json` looks the same before/after

Change-Id: I78674286e12bee16f03b002447fe70c557188814
Reviewed-on: https://gerrit.chromium.org/gerrit/50370
Reviewed-by: Liam McLoughlin <lmcloughlin@chromium.org>
Tested-by: Liam McLoughlin <lmcloughlin@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>