cgpt: export details for all partitions
Knowing how big a partition is can be useful, but it's a lot more useful
to know how much data is filled in that partition.  This way we don't
waste time writing out unused data (usually just zeros).  This makes a
bigger difference on flash based devices where erasing/writing is not
only slow, but can degrade the life of the device.

Here we take the existing WriteRootPartitionSize func and generalize it
so that it generates info for all partitions.  At runtime, we can load
the right set of data and then refer to it by number.

BUG=chromium:236941
TEST=`./build_library/cgpt.py write usb build_library/legacy_disk_layout.json f.sh` includes new funcs
TEST=`cbuildbot {amd64,arm,x86}-generic-full` works

Change-Id: I3b756e3b8f5c24097cdedf46902ae7a503c5388f
Reviewed-on: https://gerrit.chromium.org/gerrit/60119
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>