Accept a symlink'd chrome as valid for pgo_generate.
When chrome is built with USE="pgo_generate", it is installed in the stateful
partition and the usual location in rootfs is a symlink to that. This change
makes test_image_content.sh work with USE="pgo_generate" by accepting a symlink
in place of a valid chrome file.

BUG=chromium:240498
TEST=USE="pgo_generate" ./build_packages --board=$board
./build_image --board=$board --disk_layout=2gb-rootfs test

Change-Id: Ia13a011f9baaa6490e1747a1c4c860c3d8525485
Reviewed-on: https://gerrit.chromium.org/gerrit/57355
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: asharif <asharif@chromium.org>
Tested-by: asharif <asharif@chromium.org>