Modified grub.cfg to set default value based on partition priority
In this commit, template of grub.cfg is modified to set the default
parition to boot based on priority of the partition in the disk.
Grub reads the priority of partitions in the GPT disk with the help
of grub command "gptpriority". This change is required, say when
a new image is installed with image_to_live, system reboots to the
default image instead of newly installed image. This commit will
ensure the image booted is with high priority (in above case newly
installed image) instead of just booting default always. Note: This
above scenario works fine with coreboot, and the changes is to make
EFI boot behave the sameway

BUG=None
TEST=Tested by changing the parition priority manually using cgpt and
     verified that system boots to high priority image on reboot. And
     also verified image_to_live boots to newly installed image.

CQ-DEPEND=CL:I7cce1079b356e8459409113a4c17d62b25c5cd4e

Change-Id: I930d0f3f1a3f8f54edd1dce7f8259e3c261af9a1
Reviewed-on: https://gerrit.chromium.org/gerrit/58276
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Sathish Kuppuswamy <sathish.kuppuswamy@intel.com>
Commit-Queue: Josh Triplett <josh@joshtriplett.org>