security_test_image: default to using signer vboot rev
Rather than running security tests against the ToT vboot, use the same
git ref that the signer does.  This way, when a change is landed in the
vboot repo that the build logic uses but breaks with the older vboot,
we'll notice in the build bot phase rather than when the signer runs.

BUG=chromium:255529
TEST=`./security_test_image --board=stumpy` checked out the vboot ref the signer uses

Change-Id: I5e6ea9495e498c092f8ed80b81d9693dd133f034
Reviewed-on: https://gerrit.chromium.org/gerrit/60502
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>