Do not switch off serial console
Linux kernel deregisters the serial console as soon as other consoles
become available (for instance LCD). This is not always optimal, as
the kernel sometimes hangs after alternative consoles became
available, but before user logs in.

Adding keep_bootcon to the command line keeps serial console enabled
for the entire duration. Note that the messages keep coming out even
after login screen appears, but this is an inconvenience one should be
able to live with.

BUG=chromium:260683
TEST=manual
  . with this change the serial console output is not shut off mid
    kernel startup.

Change-Id: I49be4739ce54f0a86b3433b8b73dc9d9acaebed0
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62005