Remove old versions of autotest-tests and autotest-deps.
This CL is copied from:
https://gerrit.chromium.org/gerrit/#/c/61600

To speed up the build, we are splitting the autotest-tests and
autotest-deps packages into a few other packages. This may cause
conflicts so we will need to uninstall the other packages first. This
migration script will take care of the uninstallation.

BUG=chromium:238407
TEST=none

Change-Id: Idd85c84279318ce1aaad07fa31e479c732e39752
Reviewed-on: https://gerrit.chromium.org/gerrit/62387
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Dan Shi <dshi@chromium.org>