make_chroot: make bare_chroot more robust
With some distros merging /bin and /sbin and /usr, assuming the $PATH
outside of the chroot contains all the elements we need when we're in
the chroot no longer flies.  Update the bare_chroot helper so that it
always includes all the paths that matter inside the chroot.

BUG=None
TEST=`cros_sdk --chroot asdf` still works

Change-Id: Ieacf2772448b8b3803f60b78412aa5d179e6b49f
Reviewed-on: https://gerrit.chromium.org/gerrit/64336
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>