Create tool to validate files before copying them.
The intention is that this tool will be used to verify files that will
be moved from unencrypted stateful partion to encrypted stateful partion
using a validation file stored on the rootfs.

The same tool will be run by build image to create a validation file, and
again by the boot scripts at first boot.

This CL creates a tool for doing a validated copy, but does not use the
tool in any way.

BUG=chromium:269788
TEST=It passes the new unittests included.

Change-Id: I75bc66040a06bf0a4a2a8512c96e12a03e14da3c
Reviewed-on: https://gerrit.chromium.org/gerrit/65527
Reviewed-by: Will Drewry <wad@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>