Make /mnt/stateful_partition easily available for ebuild installation.
By mounting the stateful partition when images are being built, ebuilds
can easily install files to stateful by just referencing the same
directory paths that will be used after the image is booted.

Not that files in stateful will NOT be installed if the whitelist in
chromeos-install is not updated.

BUG=chromium:259592
TEST=Temp modified an ebuild to install to /mnt/stateful_partition, and
     observed that the file was present in stateful afterwards. Copied
     to USB and booted the image for real to confirm.

Change-Id: I6d1fdbe3da64b92ccc56595777e98657a4aadc13
Reviewed-on: https://gerrit.chromium.org/gerrit/65788
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>