cros_generate_breakpad_symbols: use 64bit dump_syms all the time
Upstream fixed dump_syms so that it is no longer wordsize specific.  That
means you can use the 64bit version of dump_syms on 32bit ELFs.  This is
good because the current Chrome builds are producing a debug file so big
that it's filling up the VM space of the 32bit ELF.  The 64bit one has no
such problem as it has a bigger VM space for maps.

BUG=chromium:266064
TEST=ran cros_generate_breakpad_symbols before & after change and compared breakpad/ output -- they were the same

Change-Id: Ib564fa1f3acc17fcf26ddcbaac2b46471908f41f
Reviewed-on: https://gerrit.chromium.org/gerrit/66214
Reviewed-by: Ben Chan <benchan@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>