add migration script to unmerge chromeos-chrome
For the migration hooks in chromeos-chrome (in the CQ-DEPEND'ed) CL to
run on incremental builds or on developer's desks, it is important that
when first building the CQ-DEPEND'ed CL, chromeos-chrome is re-emerged.
This should nomally happen anyway, since chromeos-chrome depends on
autotest, which is being uprevved by that CL. However, unmerging it
first is a good insurance policy against a problem in that logic causing
tree breakage.

CQ-DEPEND=CL:I836d933e8a828177304dd10a74841138a9502aff
BUG=chromium:256906
TEST=See CQ-DEPEND CL

Change-Id: I6b21f6921e5ea8704d67623c32839b2a04d54f8e
Reviewed-on: https://chromium-review.googlesource.com/172273
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>