Fix "setupTestingInterface" script interface naming
My change in https://gerrit.chromium.org/gerrit/37057 changed the
"testif" variable in the "setupTestingInterface" mod-for-test script
to include the testing WiFi interfaces.  This completely broke the
"setup-backchannel" script, since it generated an unusable interface
name.  Separate out the testing WiFi interfaces into a separate name
so as not to taint later, less obvious uses of the "testif" variable.

BUG=chromium:305821
TEST=Test the script out on a copied /etc and /sbin root filesystem:
    ROOT_FS_DIR=/tmp/root_fs_dir \
        bash -x mod_for_test_scripts/100setupTestingInterface
Run this change on a trybot

Change-Id: I31f36924a5c3892333f5d8d717400b2feef9c48e
Reviewed-on: https://chromium-review.googlesource.com/172616
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>