update_kernel: add remote_bootargs
This allows you to update the kernel on the target without having the
previous .../build/images/<board>/latest/config.txt around without losing
reference to rootfs, etc.

The downside is that the root= path and other args that are normally
dynamically built up by the vboot wrapper now are hardcoded, so in some
settings this can cause problems. Because of this, don't set it as default
(yet).

BUG=none
TEST=update_kernel with and without --remote_bootargs to make sure old
  and new behavior still works.

Change-Id: I6e04fa492d8ad15356a7522c12894c379141e251
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/173218