cros_vm_lib: cleanly shutdown KVM when possible
There are multiple ways to shut down a KVM instance, with some being cleaner
than others. In particular, shutting down a KVM instance using a signal
(SIGHUP, SIGINT, SIGTERM) does not inform the guest OS that the machine is
shutting down. Consequently, writes buffered in the guest kernel will not be
flushed to disk.

This poses a problem in cases where we want to use the disk image to
investigate a failure inside the VM. To resolve this problem, we use KVM's
monitor interface to initiate an ACPI shutdown instead. The KVM instance is
given about 30 seconds to complete ACPI shutdown and VM power-off, after
which the KVM is terminated with prejudice (SIGKILL).

BUG=chromium:311837
TEST=manual (see below)

Manual testing
--------------
$ cd <sandbox>/src/scripts  # outside chrooot
$ bin/cros_start_vm --image_path=<absolute path> --kvm_pid=/tmp/kvm.pid
$ ssh -p 9222 root@localhost date; bin/cros_stop_vm --kvm_pid=/tmp/kvm.pid
  record the time from the output of date here. e.g. "14:57:50"
$ bin/cros_start_vm --image_path=<absolute path> --kvm_pid=/tmp/kvm.pid
$ ssh -p 9222 root@localhost grep ssh /var/log/messages | grep <time>
  expect to find some matching lines, which roughly look like:
  <time> localhost sshd[3328]: Accepted publickey for root from 10.0.2.2 port 37811 ssh2
  <time> localhost sshd[3328]: Received disconnect from 10.0.2.2: 11: disconnected by user

Change-Id: I49291391c03aa508351fb63601cc7509d4f0569f
Reviewed-on: https://chromium-review.googlesource.com/177375
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>