cgpt: support percentages and negative counts in fs_blocks
For some partitions, we don't want to hand maintain the min fs size, but
just make sure that there are some number of blocks left over at the end.
To simplify that, extend the syntax so that we can say "have the fs take
up 90% of the partition" or "have the fs take up the whole partition, but
leave 10 blocks free at the end".

BUG=chromium:323593
TEST=ran cgpt.py debug output on various json files

Change-Id: I7ba979ecb267999a007f8723e58f585831927ee4
Reviewed-on: https://chromium-review.googlesource.com/178231
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>