cgpt: support stacked layouts
Currently, if a board wants to override the main disk layout, it has to
do so by duplicating the entire table.  This is a pain when the board
often times wants to tweak only one or two partitions.

Instead, allow the layout to declare a parent layout.  This will instruct
the code to load that layout first and then merge this layout with it.
The number (and depth) of parents is infinite.

This first implementation requires the parent to be a full or relative
path.  In the future we'll probably want to expand by supporting things
like portage's repo notation.

BUG=chromium:323593
TEST=ran cgpt.py debug output on various json files

Change-Id: I76c775548700f04a3a91ac7c5f2307b68ea36b0e
Reviewed-on: https://chromium-review.googlesource.com/178232
Reviewed-by: Liam McLoughlin <lmcloughlin@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>