mount / with --mount-rslave before entering chroot
When using systemd as the init system, entering the cros chroot, the
cros mounts show up outside the chroot, and they also seem to stick
around afterwards.  According to the kernel documentation
sharedsubtree.txt we want to use --make-rslave so that our process is
cloned off in a new namespace, and marked as slave.

BUG=None
TEST=Ran mount inside and outside chroot before and after the change
when using systemd as init system as well as using openrc as the init
system.

Change-Id: I9f564e63e8227bb9a2d5a425a849576d55c8da9e
Signed-off-by: Steev Klimaszewski <steev@gentoo.org>
Reviewed-on: https://chromium-review.googlesource.com/178352
Reviewed-by: Steev Klimaszewski <threeway@gmail.com>
Tested-by: Steev Klimaszewski <threeway@gmail.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>