cgpt: improve the inheritance logic to respect new base over parent override.
This CL changes the inheritance logic in two ways:
1) Instead of a e.g. new usb layout inheriting directly from the usb layout
of its parent, it instead inherits the usb layout from the inherited file +
its own base layout before applying its own changes. This means if an overlay
specifies an even smaller base partition size than in the usb layout of the
parent, the smaller base partition wins e.g.

local usb <- local base <- parent usb <- parent base.

2) Respect the superset of layout names defined by all the parent and main
config. This means if a usb layout isn't defined in a new file but in its
parent, that exists but is overriden by the base that is defined in the new
file. Similarly, if a new file defines a layout not present in its parent,
that layout inherits from the parent's base and its own base.

BUG=chromium:335269
TEST=compared beaglone and legacy configs before and after. in addition added
my own new layout in that overlay to verify both sides of part 2.

Change-Id: I17738ca934e17aa8028f86b88a700bc8cf17f73f
Reviewed-on: https://chromium-review.googlesource.com/183391
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Chris Sosa <sosa@chromium.org>