Fix issue when passing in package category
When also passing in the package category as a parameter (such as
'virtual/chromeos-bsp' rather than just 'chromeos-bsp'), the
script would throw an error when trying to create the temporary
install folder.  For the temporary folder name, normalize all '/'
to '_' in the package name.

BUG=none
TEST=`./package_dependency_runtime_stats.sh --board=beaglebone \
  virtual/chromeos-bsp` runs

Change-Id: Idcf1ca0c984fd5ff867d76984d091952e1465c1a
Reviewed-on: https://chromium-review.googlesource.com/184694
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Steve Fung <stevefung@chromium.org>
Commit-Queue: Steve Fung <stevefung@chromium.org>
Tested-by: Steve Fung <stevefung@chromium.org>