build_kernel_image: fix duplicate serial console messages
The inclusion of console=, earlyprintk=, and keep_bootcon
command line options for the serial port results in
duplicate kernel messages on the serial port. The reason is
that earlyprintk with keep_bootcon will keep that console
spitting out messages and the console associated with
console= will continue to spit out the same messages.

In order to maintain the early console from boot without
the code duplication just use earlyprintk and keep_bootcon
instead of additionally using another console= line.

BUG=None
TEST=Built and image with this change. No more duplicate kernel
     messages.

Change-Id: I50e5e8227c754142d154719a4a5b697bc9806c8d
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/184575
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>