legacy_disk_layout: bump up rootfs size on recovery and usb
"CGpt: Add 200MiB to the Rootfs" increased the filesystem size on the
rootfs of built images. For most platforms there was still room for
the verity blocks at the end of partition 3, but for parrot, which uses
bootcache, the new filesystem was too large to fit together with the verity
blocks and the 100000 bootcache blocks.

Since we always use the same bootargs independent on the media (from
installed ssd or from usb stick), we hit a situation where the
devicemapper setup tried to access beyond the end of the partition,
got an error and aborted. The result was a non-booting USB stick, and
anon-booting VM image for vmtest.

BUG=chromium:344509
TEST=build image, run image_to_vm, boot vm image. Write image to usb stick, boot that image too.

Change-Id: I5d2a5fb176069f6bfa62e4790db247fc155ca88d
Reviewed-on: https://chromium-review.googlesource.com/188183
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Olof Johansson <olofj@chromium.org>
Tested-by: Olof Johansson <olofj@chromium.org>