Ensure all changes in disk_layout are used
If there are changes in overlay disk layout but the
base layout is not changed, all other changes are ignored.
Even if there is a base layout in the overlay, partitions that are
changed but are not present in the base layout are ignored.

BUG=chromium:373840
TEST=moblab overlay:
TEST=- check that partition 5 of USB layout is set to 2048
TEST=Take moblab overlay, remove base layout.
TEST=- check that USB layout is taken into account.

Change-Id: Ib769f43fae3164c720f4e9439d98e1d214154318
Reviewed-on: https://chromium-review.googlesource.com/200083
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>