Enable licensing checks in build_packages.
Give ENABLE_LICENSING a default of Y during build_packages. This
doesn't change licensing in some host targetted ebuilds, but catches
errors in the ebuilds we ship.

Check ENABLE_LICENSING for value of Y, so that people can set it to N
if they need to turn it off.

BUG=chromium:207004
TEST=Trybots. Trybots galore.

Change-Id: I2d26b2ab10e14650f664e018d16aba6505304a76
Reviewed-on: https://chromium-review.googlesource.com/204685
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>