build_image: Remove unused gconv files when building the image.
This patch introduces a new python scripts that looks for used gconv
modules (referenced by calls to iconv_open). The python script does
that by searching for the charset string of each defined charset
on the gconv-modules file, such as 'UTF-16LE' or 'ISO-8859-1'. If the
string is found on a binary importing or defining the symbol
iconv_open, the charset is assumed to be used.

This can lend to some false positives, but most of these modules are
not used. At the moment, chrome doesn't use this infrastructure and
only a few of them are useful in our context (like UTF-16 and
UTF-32).

BUG=chromium:206111
TEST="sudo python gconv_strip.py $ROOTFS_PATH --debug --dry-run" shows
the found and removed modules.
TEST=`./build_image --board=$BOARD` shows the removed modules.
TEST=cbuildbot link-release lumpy-release daisy_spring-release

Change-Id: I57af97f696f6cd62ccf20ea4137a08e70f72d494
Reviewed-on: https://chromium-review.googlesource.com/206152
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>