gconv_strip: Improve speed using ahocorasick.
Use ahocorasick python module to improve speed on the string search
step of the gconv_script.py. Other minor changes are included in this
patch.

This patch reduces the required time from 6s to 1s on a small gizmo
image running on a Z620, assuming the image is cached in RAM.

BUG=chromium:399469
TEST=`gconv_strip.py -V --dry-run /path/to/rootfs` with and without this patch.
CQ-DEPEND=CL:210704

Change-Id: If89db9fc1c7fc6c9d5ca5bea30f8a5d2a524465c
Reviewed-on: https://chromium-review.googlesource.com/211972
Reviewed-by: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>