setup_board: fix wrapper setup
The previous change here missed escaping the $ which meant they were
expanded in the context of the script.  The resulting code never ran
(because it evaluated to ""=="1"), but it also meant it couldn't do
what it was supposed to.

While we're here, also fix a minor bug when these scripts are run
w/out HOME in the env.

BUG=None
TEST=`./setup_board --board=link --regen_configs` and read the generated wrappers

Change-Id: Idbbb0aae4089292a77991b506e8d6612d74b3010
Reviewed-on: https://chromium-review.googlesource.com/212222
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>