cgpt: disk_layout: Specify where primary entry array is
This CL passes a "-p" parameter to `cgpt create` to pad the primary GPT
header and its entries. This allows us to place the primary partition
array at an arbitrary location. The location is controlled by the value
of "primary_entry_array_lba" in the "metadata" section of the disk layout.

This is required because some board only boots from a specific block on
disk. In this case, the Freescale board only boots from the third block.

BUG=chromium:406432
TEST=Add "primary_entry_array_lba" value to "metadata" in disk_layout.json,
     then build_image. Make sure `cgpt show` the image with padding.
CQ-DEPEND=CL:213861

Change-Id: Idf9ebd1d1bcd68f41d7d1ef27565f0205cb6e866
Reviewed-on: https://chromium-review.googlesource.com/214318
Reviewed-by: Nam Nguyen <namnguyen@chromium.org>
Commit-Queue: Nam Nguyen <namnguyen@chromium.org>
Tested-by: Nam Nguyen <namnguyen@chromium.org>