Get rid of the --usb_disk option.
This option was passed around a lot, but was only actually used in
create_legacy_bootloader_templates.sh. No existing code used that option, and
that script is in build_library which implies that there shouldn't be external
callers. All other uses just pass around the option, although never actually
passing it to the only consumer. One use was even in a script which hadn't
defined the flag which presumably made it expand to an empty string.

BUG=chromium:403031
TEST=Ran a butterfly-paladin tryjob with --hwtest.

Change-Id: Ia9662ac2d6f164090369d16539fe25bf09b832ba
Reviewed-on: https://chromium-review.googlesource.com/216053
Reviewed-by: Will Drewry <wad@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>