image_to_vm.sh: Set up the kernel partitions after make_image_bootable.
Calling make_image_bootable is difficult to avoid since we modify the rootfs
and create a new GPT whose partitions have different UUIDs. Unfortunately,
that script recreates the kernel partitions, and signs the one in slot A as
a recovery kernel instead of as a normal kernel which we want.

This change stops copying the kernel partitions from the old image to the new
since they're going to be replaced anyway, and copies the kernel from slot B
to slot A.

We use losetup instead of the mechanisms used elsewhere in the script because
it's a lot easier to use. It would be nice to retrofit those other parts to
use losetup to be simpler, more consistent, and more performant.

BUG=chromium:403031
TEST=Ran a tryjob on butterfly-paladin with --hwtest.

Change-Id: Ifcaa1079962dbe9860e11051208434d4febae74c
Reviewed-on: https://chromium-review.googlesource.com/216372
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>