setup_board: ENABLE_LICENSING=Y in the emerge wrapper.
In the emerge wrapper for all boards, enable licensing checks. This
should mean that licensing is checked for everything that gets
installed on a board, but we won't get the errors for licensing
problems with host targetted packages (there are some, not yet fixed).

BUG=chromium:413937
TEST=cbuildbot --remote -g 217974 link-release \
  `cbuildbot --list | grep release-group | grep -v freon | cut -f1 -d' '`

Change-Id: I843bd942eea57fae76999457d85d52f903e7d4db
Reviewed-on: https://chromium-review.googlesource.com/217974
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>